requirements.txt
PyPDF2
pyspellchecker

# Обработка файлов и тесты
ruff
pytest
pre-commit